requirements.txt
pygame
PyOpenGL
numpy
pyttsx3